Rearranged classes to follow the MVVM architecture
