+++ TaleKeeper/app/src/main/java/com/carlosreads/talekeeper/viewmodels/LibraryViewModel.java
@@ -1,16 +1,16 @@
-package com.carlosreads.talekeeper.ui.notifications;
+package com.carlosreads.talekeeper.viewmodels;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class ProfileViewModel extends ViewModel {
+public class LibraryViewModel extends ViewModel {
 
     private final MutableLiveData<String> mText;
 
-    public ProfileViewModel() {
+    public LibraryViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is notifications fragment");
+        mText.setValue("This is home fragment");
     }
 
     public LiveData<String> getText() {
