+++ TaleKeeper/app/src/main/java/com/carlosreads/talekeeper/viewmodels/ProfileViewModel.java
@@ -1,16 +1,16 @@
-package com.carlosreads.talekeeper.ui.home;
+package com.carlosreads.talekeeper.viewmodels;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class LibraryViewModel extends ViewModel {
+public class ProfileViewModel extends ViewModel {
 
     private final MutableLiveData<String> mText;
 
-    public LibraryViewModel() {
+    public ProfileViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is home fragment");
+        mText.setValue("This is notifications fragment");
     }
 
     public LiveData<String> getText() {
