+++ TaleKeeper/app/src/main/res/layout/fragment_profile.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.notifications.ProfileFragment">
+    tools:context=".views.ProfileFragment">
 
     <TextView
         android:id="@+id/text_notifications"
