+++ TaleKeeper/app/src/main/res/navigation/mobile_navigation.xml
@@ -7,19 +7,19 @@
 
     <fragment
         android:id="@+id/navigation_library"
-        android:name="com.carlosreads.talekeeper.ui.home.LibraryFragment"
+        android:name="com.carlosreads.talekeeper.views.LibraryFragment"
         android:label="@string/library"
         tools:layout="@layout/fragment_library" />
 
     <fragment
         android:id="@+id/navigation_discover"
-        android:name="com.carlosreads.talekeeper.ui.dashboard.DiscoverFragment"
+        android:name="com.carlosreads.talekeeper.views.DiscoverFragment"
         android:label="@string/discover"
         tools:layout="@layout/fragment_discover" />
 
     <fragment
         android:id="@+id/navigation_profile"
-        android:name="com.carlosreads.talekeeper.ui.notifications.ProfileFragment"
+        android:name="com.carlosreads.talekeeper.views.ProfileFragment"
         android:label="@string/profile"
         tools:layout="@layout/fragment_profile" />
 </navigation>
